Fix incorrect error displaying logic
Previously the error would be displayed whenever anything that was not
an identifier would be given. This was because the scanIdentifier-
method was the last scanning method, and it was presumed that if the
given character was not a beginning of an identifier, an error was
encountered. However, the inputted character was given to each scanning
method whether they recognized it. Thus, the scanIdentifier-
method would get all characters and not report and error for only the
ones that could start an identifier.