Add associativity and precedence to expression grammar
This new grammar definition produces syntax trees with correct precedence
and associativity. Most expressions are defined as sequences of operations
instead of using left-recursion. This is because of the parsing technique
we are going to use; it does not handle left recursion well. This syntax
more accurately represent the actual parsing.